Move incompatible message warning into properties
The maven-hpi-plugin usage inside the build section of the pom file is
allowed but the simpler property setting is listed as preferred in the
Jenkins wiki.

Also has the benefit of removing a line with trailing white space from
the pom.  Unclear why tidy:pom did not flag that trailing white space
as a violation.