Do not run parallel tests
Parallel tests hang at exit on my 2+ year old FreeNAS mini XL with
its 4 core (8 thread) low end processor. The performance gain from
threaded execution of the relatively few tests in the plugin are
not worth fighting with Maven and Jenkins and more to determine
which component is failing to do which part of the work for
multi-process and multi-threaded tests.